+++ eslint.config.js
@@ -3,8 +3,8 @@
 const globals = require("globals");
 const js = require("@eslint/js");
 
-const nodePlugin = require("eslint-plugin-n").default;
-const jestPlugin = require("eslint-plugin-jest");
+const nodePlugin = require("eslint-plugin-n");
+const vitestPlugin = require("@vitest/eslint-plugin");
 const configPrettier = require("eslint-config-prettier");
 
 // https://eslint.org/docs/latest/use/configure/configuration-files
@@ -15,7 +15,8 @@ module.exports = [
   },
   js.configs.recommended,
   nodePlugin.configs["flat/recommended-script"],
-  jestPlugin.configs["flat/recommended"],
+  vitestPlugin.configs.recommended,
+  vitestPlugin.configs.env,
   configPrettier,
   {
     languageOptions: {
@@ -55,8 +56,20 @@ module.exports = [
       "n/no-unpublished-require": ["off"],
       "n/no-unsupported-features/node-builtins": ["off"],
       "n/no-deprecated-api": ["off"],
-      "jest/no-conditional-expect": ["off"],
-      "jest/no-disabled-tests": ["off"],
+      "vitest/no-conditional-expect": ["off"],
+      "vitest/no-disabled-tests": ["off"],
+    },
+  },
+  {
+    files: ["bin/templates/node/**/*.js"],
+    languageOptions: {
+      sourceType: "module",
+    },
+    rules: {
+      strict: ["off"],
+      "n/no-extraneous-import": ["off"],
+      "n/no-missing-import": ["off"],
+      "n/no-unpublished-import": ["off"],
     },
   },
 ];