+++ java/project/srv/src/test/java/com/github/capjscommunity/sapafcsdk/test/MockMvcSecurityTestAutoConfiguration.java
@@ -1,0 +1,15 @@
+package com.github.capjscommunity.sapafcsdk.test;
+
+import org.springframework.boot.autoconfigure.AutoConfiguration;
+import org.springframework.boot.webmvc.test.autoconfigure.MockMvcBuilderCustomizer;
+import org.springframework.context.annotation.Bean;
+import org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers;
+
+@AutoConfiguration
+public class MockMvcSecurityTestAutoConfiguration {
+
+  @Bean
+  public MockMvcBuilderCustomizer mockMvcSecurityBuilderCustomizer() {
+    return builder -> builder.apply(SecurityMockMvcConfigurers.springSecurity());
+  }
+}